pyproject.toml
[tool.poetry]
name = "umuttepe-hava-botu"
version = "1.0.0"
description = "Twitter bot that tweets about Umuttepe weather conditions."
authors = ["Sinan Bekar <sinanbekar.work@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "3.10.0"
requests = "2.22.0"
beautifulsoup4 = "4.10.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
azure-functions = "^1.11.3"
python-dotenv = "0.19.0"

[tool.poetry.scripts]
start = "umuttepe_hava_botu:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
